将route信息单独放在一个js文件里方便管理，也符合编程习惯
